introduce the relation plugin
There are many relations between K8s resources, like owner
reference, label selector, etc...
It becomes challenging for dashboards to manage these relations
case by case, so we implemented the relation plugin to handle
this in the API provider.
When we store data in the global store, we will process each item
and append relations into the metadata field.
When we apply data back to the backend, we will restore each item
to ensure the fields we add will not be sent to the API server.

There are some potentials to make the plugin system more clear,
for example, use pre/post hooks to organize the code. But now,
we will keep the code simple, and go back to review the design
when we have more plugins.